Tests/LibWeb/Text/input/HTML/custom-element-parser-insertion-reactions.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    let reactions = [];
    customElements.define("parser-insertion-reaction", class extends HTMLElement {
        constructor() {
            super();
            reactions.push("constructor");
        }

        connectedCallback() {
            reactions.push("connected");
        }
    });
</script>
<parser-insertion-reaction></parser-insertion-reaction>
<script>
    let reactionsAfterParserInsertion = reactions.join(",");
    test(() => {
        println(`after parser insertion: ${reactionsAfterParserInsertion}`);
        println(`at DOMContentLoaded: ${reactions.join(",")}`);
    });
</script>
