Tests/LibWeb/Text/expected/HTML/custom-element-parser-insertion-reactions.txt

after parser insertion: constructor,connected
at DOMContentLoaded: constructor,connected
